+++ src/languages.ts
@@ -31,6 +31,7 @@ import SignatureManager from './provider/signatureManager'
 import TypeDefinitionManager from './provider/typeDefinitionManager'
 import TypeHierarchyManager, { TypeHierarchyItemWithSource } from './provider/typeHierarchyManager'
 import WorkspaceSymbolManager from './provider/workspaceSymbolsManager'
+import type { Sources } from './sources'
 import { ExtendedCodeAction, LocationWithTarget, ProviderName, TextDocumentMatch } from './types'
 import { disposeAll } from './util'
 
@@ -101,7 +102,7 @@ class Languages {
     allCommitCharacters?: string[]
   ): Disposable {
     selector = typeof selector == 'string' ? [{ language: selector }] : selector
-    let sources = require('./sources/index').default
+    let sources = require('./sources/index').default as Sources
     sources.removeSource(name)
     return sources.createLanguageSource(name, shortcut, selector, provider, triggerCharacters, priority, allCommitCharacters)
   }